Move api code to apis/verrazzano folder
