+++ operator/controllers/upgrade_reconciler.go
@@ -5,7 +5,8 @@ package controllers
 
 import (
 	"fmt"
-	installv1alpha1 "github.com/verrazzano/verrazzano/operator/api/v1alpha1"
+
+	installv1alpha1 "github.com/verrazzano/verrazzano/operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/operator/internal/component"
 	"go.uber.org/zap"
 	ctrl "sigs.k8s.io/controller-runtime"
