+++ operator/controllers/upgrade_reconciler_test.go
@@ -6,15 +6,16 @@ package controllers
 import (
 	"context"
 	"errors"
-	"github.com/verrazzano/verrazzano/operator/internal/util/helm"
 	"os/exec"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	"testing"
 	"time"
 
+	"github.com/verrazzano/verrazzano/operator/internal/util/helm"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
-	vzapi "github.com/verrazzano/verrazzano/operator/api/v1alpha1"
+	vzapi "github.com/verrazzano/verrazzano/operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/operator/mocks"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
